added support for git branch

+++ bashrc
@@ -8,19 +8,35 @@ export PATH=$PATH:/Library/TeX/texbin:/usr/local/bin:~/bin:~/lib:~/.minishift/ca
 #################################################################
 
 ###########
-#less
+# git branch for PS1
+function gb() {
+    if command -v git >/dev/null 2>&1; then
+        echo -n '(' && git branch 2>/dev/null | grep '^*' | colrm 1 2 | tr -d '\n' && echo  -n ')'
+    fi
+}
+readonly -f gb
+[ "$?" -eq "0" ] || return $?
+
+function git_branch() {
+    gb | sed 's/()//'
+}
+readonly -f git_branch
+[ "$?" -eq "0" ] || return $?
+
+###########
+# less
 alias less='less --RAW-CONTROL-CHARS'
 
 ###########
-#rm
+# rm
 alias rm='rm -i'
 
 ###########
-#rm
+# mv
 alias mv='mv -i'
 
 ###########
-#ls
+# ls
 export LS_OPTS='-Flh'
 
 #if color is available use it
@@ -33,16 +49,16 @@ alias lll='ls'
 alias lla='ls -a'
 
 ###########
-#grep
-#if color is available use it
-#OS X does not support --color=auto
+# grep
+# if color is available use it
+# macOS does not support --color=auto
 export GREP_OPTS=' -in' 
 export GREP_COLOR='1;32'
 
 grep --color=auto > /dev/null 2>&1 && alias grep='grep ${GREP_OPTS} --color=auto' || alias grep='grep ${GREP_OPTS}'
 
 ###########
-#text editor
+# text editor
 alias bashrc='vim ~/.bashrc'
 
 #################################################################
@@ -82,4 +98,4 @@ export CLICOLOR=1
 UC=$COLOR_LIGHT_GREEN               # user's color
 [ $UID -eq "0" ] && UC=$COLOR_RED   # root's color
 
-PS1="\n\[${COLOR_LIGHT_GRAY}\]\t \[${UC}\]\u \[${COLOR_LIGHT_CYAN}\]\h \[${COLOR_YELLOW}\]\${PWD} \n\[${COLOR_LIGHT_GREEN}\]→\[${COLOR_NC}\] "
+PS1="\n\[${COLOR_LIGHT_GRAY}\]\t \[${UC}\]\u \[${COLOR_LIGHT_CYAN}\]\h \[${COLOR_YELLOW}\]\${PWD} \[${COLOR_RED}\]\$(git_branch) \n\[${COLOR_LIGHT_GREEN}\]→\[${COLOR_NC}\] "
